Create listeners in the final initialization step (#530)
